enabled nullable

+++ Project.Core/Models/Room.cs
@@ -11,8 +11,8 @@ namespace Project.Core.Models
         public string RoomName { get; set; }
         public int Capacity { get; set; }
         public int CurrentCapacity { get; set; }
-        public double Price { get; set; }
-        public double Debt { get; set; }
+        public decimal Price { get; set; }
+        public decimal Debt { get; set; }
         public string RoomType { get; set; }
         public bool Lack { get; set; }
         public string LackDetail { get; set; }
